現在時刻の部分をコンポーネント分割
